Trait bits::generic::BitOps

pub trait BitOps: Int {
    fn b_and_not(self, rhs: Self) -> Self { ... }
    fn b_get_block_from(self, start: usize, len: usize) -> Self { ... }
    fn b_get_block(self) -> Self { ... }
    fn b_gather(self, mask: Self) -> Self where Self: Neg<Output=Self> { ... }
    fn b_get_border(self) -> Self { ... }
    fn b_get_interior(self) -> Self { ... }
    fn b_get_singletons(self) -> Self { ... }
    fn b_get_singletons_xi(self) -> Self { ... }
    fn b_scatter(self, mask: Self) -> Self where Self: Neg<Output=Self> { ... }
    fn b_zero_high_from_pos(self, index: usize) -> Self { ... }
    fn i_count_total(self) -> u32 { ... }
    fn i_fill_upto(self) -> Self { ... }
    fn i_get(self) -> Self where Self: Neg<Output=Self> { ... }
    fn i_get_block(self) -> Self { ... }
    fn i_get_block_border(self) -> Self { ... }
    fn i_get_block_border_low(self) -> Self { ... }
    fn i_get_block_border_high(self) -> Self { ... }
    fn i_get_border(self) -> Self { ... }
    fn i_get_border_high(self) -> Self { ... }
    fn i_get_border_low(self) -> Self { ... }
    fn i_get_interior(self) -> Self { ... }
    fn i_get_invert(self) -> Self { ... }
    fn i_get_singletons(self) -> Self { ... }
    fn i_get_singletons_xi(self) -> Self { ... }
    fn i_mask_upto_invert(self) -> Self { ... }
    fn i_mask_upto(self) -> Self { ... }
    fn i_flip(self) -> Self { ... }
    fn o_count_leading(self) -> u32 { ... }
    fn o_count_total(self) -> u32 { ... }
    fn o_count_trailing(self) -> u32 { ... }
    fn o_flip(self) -> Self { ... }
    fn o_fill_upto(self) -> Self { ... }
    fn o_get(self) -> Self { ... }
    fn o_get_invert(self) -> Self { ... }
    fn o_get_singletons(self) -> Self { ... }
    fn o_get_singletons_xi(self) -> Self { ... }
    fn o_mask_upto(self) -> Self { ... }
    fn o_mask_upto_invert(self) -> Self { ... }
}

Provided Methods

fn b_and_not(self, rhs: Self) -> Self

Logical and not

fn b_get_block_from(self, start: usize, len: usize) -> Self

Get bit block

fn b_get_block(self) -> Self

Get bits included in blocks k > 1 count bits of the same parity

fn b_gather(self, mask: Self) -> Self where Self: Neg<Output=Self>

Gather bits

fn b_get_border(self) -> Self

Return 0s and 1s on block transitions within outer 0s

fn b_get_interior(self) -> Self

Get bits bounded on the left and right by bits of the same parity

fn b_get_singletons(self) -> Self

Get singleton 0s and 1s within outer 0s

fn b_get_singletons_xi(self) -> Self

Get singleton 0s and 1s

fn b_scatter(self, mask: Self) -> Self where Self: Neg<Output=Self>

Scatter bits

fn b_zero_high_from_pos(self, index: usize) -> Self

Zero high bits from specified position

fn i_count_total(self) -> u32

Count total 1s; population count

fn i_fill_upto(self) -> Self

Fill up to lowest 1

fn i_get(self) -> Self where Self: Neg<Output=Self>

Get lowest 1

fn i_get_block(self) -> Self

Get 1s included in blocks k > 1 count 1s

fn i_get_block_border(self) -> Self

Get 1s bordering blocks of k > 1 count 1s

fn i_get_block_border_low(self) -> Self

Get 1s to the left of blocks of k > 1 count 1s

fn i_get_block_border_high(self) -> Self

Get 1s to the right of blocks of k > 1 count 1s

fn i_get_border(self) -> Self

Get 1s bordering 0s

fn i_get_border_high(self) -> Self

Get 1s to the right of 0s

fn i_get_border_low(self) -> Self

Get 1s to the left of 0s

fn i_get_interior(self) -> Self

Get 1s bounded by 1s on the left and right

fn i_get_invert(self) -> Self

Get lowest 1 and complement word

fn i_get_singletons(self) -> Self

Get singleton 1s within outer 0s

fn i_get_singletons_xi(self) -> Self

Get singleton 1s

fn i_mask_upto_invert(self) -> Self

Mask trailing 1s and complement word

fn i_mask_upto(self) -> Self

Mask up to lowest 1

fn i_flip(self) -> Self

Set lowest 1 to 0

fn o_count_leading(self) -> u32

Count leading 0s

fn o_count_total(self) -> u32

Count total 0s

fn o_count_trailing(self) -> u32

Count trailing 0s

fn o_flip(self) -> Self

Set lowest 0 to 1

fn o_fill_upto(self) -> Self

Fill up to lowest 0

fn o_get(self) -> Self

Get lowest 0

fn o_get_invert(self) -> Self

Get lowest 0 and complement word

fn o_get_singletons(self) -> Self

Get singleton 0s within outer 0s

fn o_get_singletons_xi(self) -> Self

Get singleton 0s

fn o_mask_upto(self) -> Self

Mask up to lowest 0

fn o_mask_upto_invert(self) -> Self

Mask trailing 0s